Redesign API of runner package
